app/requirements.txt
celery==5.1.2
Flask==2.0.1
Jinja2==3.0.1
pillow
redis

